requirements.txt
# Stage 1: Training
tensorflow>=2.10.0
keras>=2.10.0
scikit-learn>=1.2.0
numpy>=1.20.0
matplotlib>=3.5.0
joblib>=1.2.0
umap-learn>=0.5.0

# Stage 2: Prediction
# (same as above)

# Stage 3: Web Interface
flask>=2.2.0
werkzeug>=2.2.0

# Optional (for better development)
python-dotenv>=0.21.0
